Maximum Likelihood Expectation Maximisation Algorithm with Regularised Point Spread Function
The algorithm is a version of Maxmimum Likelihood expectation with regularised Point spread function to remove spectral artefacts known as 'ghosts' that occurs in CT scanners due to limitation of number of projections extracted. The algorithm is tested with an input of an image which is transformed in to projections and then its filtered followed by backprojection to obtain the image.